requirements_minimal.txt
# Minimal requirements for deployment
# This file contains only essential dependencies for running the app

# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# LLM Integration
openai==0.28.0
anthropic>=0.7.0

# Analysis
scikit-learn>=1.3.0
networkx>=3.1

# Visualization
plotly>=5.17.0

# Configuration
pyyaml>=6.0
click>=8.1.0